+++ tests/test_vector.py
@@ -9,8 +9,8 @@ def test_basic():
     assert str(vector()) == "vector([], [])"
     vec = vector(range(3))
     assert len(vec) == 3
-    assert np.array_equal(vec.keys(), np.array([2, 1, 0]))
-    assert np.array_equal(vec.values(), np.array([1, 1, 1]))
+    assert np.array_equal(vec.toarrays()[0], [2, 1, 0])
+    assert np.array_equal(vec, [1, 1, 1])
     assert np.array(vec).dtype == "float64"
     assert np.array(vec, dtype=int).dtype == "int64"
     assert vec[0] == 1.0
@@ -28,10 +28,11 @@ def test_basic():
     vec.clear()
     assert not vec
     vec = vector(vector(range(3), 2.0))
-    assert np.array_equal(vec.values(), np.array([2.0, 2.0, 2.0]))
+    assert np.array_equal(vec, [2.0, 2.0, 2.0])
     vec = vector({0: 0.0, 1: 1.0})
-    assert set(vec.values()) == {0.0, 1.0, 0.0}
-    vec.update(vector([1, 2]).keys())
+    with pytest.warns(DeprecationWarning):
+        assert set(vec.values()) == {0.0, 1.0}
+    vec.update(vector([1, 2]))
     assert dict(vec.items()) == {0: 0.0, 1: 2.0, 2: 1.0}
     assert vector(np.array([False]), np.array(["0"]))
     assert vector(indices([0]))
@@ -73,7 +74,8 @@ def test_cmp():
 
 def test_dense():
     vec = vector.fromdense(range(4))
-    assert dict(vec) == {1: 1, 2: 2, 3: 3}
+    with pytest.warns(DeprecationWarning):
+        assert dict(vec) == {1: 1, 2: 2, 3: 3}
     assert np.array_equal(vec.todense(), np.arange(4, dtype=float))
     assert list(vec.todense(5)) == [0, 1, 2, 3, 0]